AuctionHouse (Market Manager v2.1.8)

Library representing the interface for the auction house. Responsible for making calls and decoding the answers from the auction house into a format the manager understands.

Functions

delete_order(order_id)

@spec delete_order(Shared.Data.PlacedOrder.t()) ::
  AuctionHouse.Type.delete_order_response()

Deletes an order from warframe market.

Example:

> placed_order = PlacedOrder.new(%{
    "item_id" => "54e644ffe779897594fa68cd",
    "order_id" => "626127cbc984ac033cd2bbd2"
})

> AuctionHouse.delete_order(placed_order)
:ok

> AuctionHouse.delete_order(placed_order)
{:error, :reason, order_id}

get_all_orders(item_name)

@spec get_all_orders(AuctionHouse.Type.item_name()) ::
  AuctionHouse.Type.get_all_orders_response()

Gets all warframe market orders for the item with the given name. The item's name is in human readable format. This function also converts the name into a format that the external party can understand.

Example:

> item_name = "Despoil"

> AuctionHouse.get_all_orders(item_name)
{:ok, [
  %Shared.Data.OrderInfo{
        "visible" => true,
        "order_type" => "sell",
        "platform" => "pc",
        "platinum" => 20,
        "user" => %Shared.Data.OrderInfo.User{
          "ingame_name" => "user_name_1",
          "status" => "ingame"
        }
      }
    ]
}

> AuctionHouse.get_all_orders(item_name)
{:error, :reason, item_name}

login(credentials)

@spec login(Shared.Data.Credentials.t()) :: AuctionHouse.Type.login_response()

Stores the user's credentials and authenticates with the auction house to make requests. Must be invoked every time the application is launched. It also performs the necessary steps for authorization. Returns user information.

Example:

> alias Shared.Data.{Authorization, Credentials, User}
> credentials = Credentials.new("the_username", "the_password")

> AuctionHouse.login(credentials)
{:ok,
  {
    %Authorization{cookie: "a_cookie", token: "a_token"},
    %User{patreon?: false, ingame_name: "fl4m3"}
  }
}

> AuctionHouse.login(credentials)
{:error, :reason, credentials}

logout()

@spec logout() :: AuctionHouse.Type.logout_response()

Deletes the current session and user data from the this application. Does not interact with the External AuctionHouse, this is a local operation only.

Example:

> AuctionHouse.logout()
:ok

place_order(order)

@spec place_order(Shared.Data.Order.t()) :: AuctionHouse.Type.place_order_response()

Places an order in warframe market.

Example:

> alias Shared.Data.{Order, PlacedOrder}
> order = Order.new(%{
  "item_id" => "54e644ffe779897594fa68cd",
  "mod_rank" => 0,
  "order_type" => "sell",
  "platinum" => 20,
  "quantity" => 1
})

> AuctionHouse.place_order(order)
{:ok,
  %PlacedOrder{
    item_id: "54e644ffe779897594fa68cd",
    order_id: "626127cbc984ac033cd2bbd2"
  }
}

> AuctionHouse.place_order(order)
{:error, :reason, order}

recover_login(auth, user)

@spec recover_login(Shared.Data.Authorization.t(), Shared.Data.User.t()) ::
  AuctionHouse.Type.recover_login_response()

Feeds the authorization information directly to the AuctionHouse. Used when the login data is being recovered from a past login. Will only fail if this service is down.

Example:

> alias Shared.Data.{Authorization, User}
> auth = Authorization.new("a_cookie", "a_token")
> user = User.new("fl4m3", false)

> AuctionHouse.recover_login(auth, user)
:ok