AuctionHouse.Impl.Settings (Market Manager v2.1.5)

Holds configurations that do not depend on the environment. 12 factor app standard states that only configurations that depend on the environment should be configurable in config files (config folder).

Since these variables are all compile time, no matter the environment they are used for, this module will hold them.

Functions

requests_per_second()

Specs

requests_per_second() :: non_neg_integer()

Returns how many requests per second the throttling queue will make.

requests_queue()

Specs

requests_queue() :: atom()

Returns the name of the throttling queue used to make requests to warframe.market.