AuctionHouse.Impl.UseCase.Data.Metadata (Market Manager v2.1.9)

Data about the Request or Response being made. Includes information important for 3rd party modules to use but that is not related to the arguments of the flow itself.

Used mostly to notify interested parties of the results of the flow. When the flow is done, the "send?" field should be set to true so the handler know it has to send the results back to the interested parties.

Types

notify()

@type notify() :: [pid()]

operation()

@type operation() :: atom()

send?()

@type send?() :: boolean()

t()

@type t() :: %AuctionHouse.Impl.UseCase.Data.Metadata{
  notify: [pid()],
  operation: atom(),
  send?: boolean()
}

Represents metadata about the request/response.

Functions

mark_to_send(meta)

@spec mark_to_send(t()) :: t()

new(operation, notify, send? \\ false)

@spec new(operation(), notify(), send?()) :: t()