Market Manager v1.0.0 Cli

synopsis: Manages sell orders in warframe.market. usage: $ ./market_manager {options} example: ./market_manager --action=activate --syndicates=new_loka,red_veil --strategy=equal_to_lowest options: --action=activate|deactivate

Can be either 'activate' or 'deactivate'. Activating a syndicate means
placing a sell order on warframe.market for each item the syndicate has
that is in the *products.json* file. Deactivating a syndicate deletes all
orders in warframe.market from the given syndicate.

--syndicates=syndicate1,syndicate2

Syndicates to be affected by the action.

--strategy=top_five_average|top_three_average|equal_to_lowest|lowest_minus_one

The strategy used by the price analyser to calculate the price at which
your items should be sold.

Types

args()

Specs

args() :: [String.t()]

dependencies()

Specs

dependencies() :: %{manager: module()}

Functions

main(args, deps \\ %{manager: Manager})

Specs

main(args(), dependencies()) :: any()

Receives the input from the user, parses it and sends it to the Manager. Returns whatever response the Manager gave or an error message if the input was malformed.

Can be invoked with ["-h"] to see the help logs.