Manager (Market Manager v2.1.3)

MarketManager is an application that allows you to make batch requests to warframe.market. This is the entrypoint of everything. If you have a module and you need to talk to MarketManager, this is who you call, the public API.

Functions

activate(syndicate, strategy)

Specs

activate(Manager.Type.syndicate(), Manager.Type.strategy()) ::
  Manager.Type.activate_response()

Activates a syndicate in warframe.market. Activating a syndicate means you put on sell all the mods the syndicate has with that are in the products.json file. The price of each mod will be calculated via a PriceAnalyst depending on which strategy you choose.

Example:

> MarketManager.activate("simaris", :lowest_minus_one)
{:ok, :success}

authenticate(credentials)

Specs

authenticate(Manager.Type.credentials()) :: Manager.Type.authenticate_response()

Saves the login information used in all requests. Required parameters are:

token: the xrfc-token used to send requests

cookie: the cookie used to id the user

Performs validation on the given information.

Example:

> MarketManager.authenticate(%{"token" => "abc", "cookie" => "123"})
{:ok, %{"token" => "abc", "cookie" => "123"}}

> MarketManager.authenticate(%{"token" => "abc"})
{:error, :unable_to_save_authentication, {:missing_mandatory_keys, ["cookie"], %{"token" => "abc"}}}

> MarketManager.authenticate(%{"token" => "abc", "cookie" => "123"})
{:error, :unable_to_save_authentication, {:enoent, %{"token" => "abc", "cookie" => "123"}}}

deactivate(syndicate)

Specs

deactivate(Manager.Type.syndicate()) :: Manager.Type.deactivate_response()

Deactivates a syndicate in warframe.market. Deactivating a syndicate means you delete all orders you have placed before that belong to the given syndicate.

Example:

> MarketManager.deactivate("simaris")
{:ok, :success}

valid_action?(action)

Specs

valid_action?(String.t()) :: boolean()

Returns true if the given action is valid, false otherwise.

Example:

> MarketManager.valid_action?("bananas")
false
> MarketManager.valid_action?("activate")
true

valid_strategy?(strategy)

Specs

valid_strategy?(String.t()) :: boolean()

Returns true if the given strategy is valid, false otherwise.

Example:

> MarketManager.valid_strategy?("bananas")
false
> MarketManager.valid_strategy?("equal_to_lowest")
true

valid_syndicate?(syndicate)

Specs

valid_syndicate?(Manager.Type.syndicate()) ::
  Store.Type.syndicate_exists_response()

Returns true if the given syndicate is valid, false otherwise. A syndicate is considered to be valid if it has an entry in the products.json file, even if that entry is empty. Returns error if an error occurs, like for example the products.json file not existing.

Example:

> MarketManager.valid_syndicate?("bananas")
{:ok, false}

> MarketManager.valid_syndicate?("red_veil")
{:ok, true}

> MarketManager.valid_syndicate?("red_veil") # products.json not found
{:error, :enoent}