Manager.Impl.Interpreter (Market Manager v2.1.5)

Core of the manager, where all the logic and communication with outer layers is. Currently, it works more like a bridge between the different ports of the application and manages data between them.

activate(syndicate, strategy, handle, deps \\ [store: Store, auction_house: AuctionHouse])

@spec activate(
  Shared.Data.Syndicate.t(),
  Shared.Data.Strategy.t(),
  Manager.Type.handle(),
  Manager.Type.dependencies()
) :: Manager.Type.activate_response()

active_syndicates(list \\ [store: Store, auction_house: AuctionHouse])

@spec active_syndicates(Manager.Type.dependencies()) ::
  Manager.Type.active_syndicates_response()

deactivate(syndicate, handle, deps \\ [store: Store, auction_house: AuctionHouse])

@spec deactivate(
  Shared.Data.Syndicate.t(),
  Manager.Type.handle(),
  Manager.Type.dependencies()
) ::
  Manager.Type.deactivate_response()

login(credentials, keep_logged_in, handle, deps \\ [store: Store, auction_house: AuctionHouse])

@spec login(
  Shared.Data.Credentials.t(),
  keep_logged_in :: boolean(),
  Manager.Type.handle(),
  Manager.Type.dependencies()
) :: Manager.Type.login_response()

logout(list \\ [store: Store, auction_house: AuctionHouse])

@spec logout(Manager.Type.dependencies()) :: Manager.Type.logout_response()

recover_login(deps \\ [store: Store, auction_house: AuctionHouse])

@spec recover_login(Manager.Type.dependencies()) ::
  Manager.Type.recover_login_response()

strategies()

@spec strategies() :: Manager.Type.strategies_response()

syndicates(list \\ [store: Store, auction_house: AuctionHouse])

@spec syndicates(Manager.Type.dependencies()) :: Manager.Type.syndicates_response()