Manager.Impl.PriceAnalyst (Market Manager v2.1.8)

Strategies calculate the optimum price for you to sell an item.

There are several strategies, some focus more on selling fast, while others on getting more profit.

The strategy to use is selected at runtime. See the Manager.Impl.Strategy for more information on this.

Functions

calculate_price(product, all_orders, strategy)

@spec calculate_price(
  Shared.Data.Product.t(),
  [Shared.Data.OrderInfo.t()],
  Shared.Data.Strategy.t()
) ::
  pos_integer()

list_strategies()

@spec list_strategies() :: {:ok, [Shared.Data.Strategy.t()]} | {:error, any()}