Manager.PriceAnalyst (Market Manager v2.1.3)

Contains the formulas and calculations for all the strategies. Strategies calculate the optimum price for you to sell an item. There are several stretagies, some focus more on selling fast, while others on getting more profit.

Functions

calculate_price(product, all_orders, strategy)

Specs

calculate_price(
  Store.Type.product(),
  [AuctionHouse.Type.order_info()],
  Manager.Type.strategy()
) ::
  non_neg_integer()

valid_strategy?(strategy)

Specs

valid_strategy?(String.t()) :: boolean()