Manager.Runtime.Worker (Market Manager v2.1.7)

Process responsible for doing asynchronous tasks. The Worker is supervised by the Server and is restarted should something fail. It communicates with the calling client via send/2, since we don't know if the client is a GenServer or not. Should this change (and we are sure the client is a GenServer) then GenServer.cast should be used instead.

Types

state()

@type state() :: keyword(module())

Functions

activate(syndicate, strategy)

@spec activate(Shared.Data.Syndicate.t(), Shared.Data.Strategy.t()) :: :ok

active_syndicates()

@spec active_syndicates() :: Manager.Type.active_syndicates_response()

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

deactivate(syndicate)

@spec deactivate(Shared.Data.Syndicate.t()) :: :ok

login(credentials, keep_logged_in)

@spec login(Shared.Data.Credentials.t(), keep_logged_in :: boolean()) ::
  Manager.Type.login_response()

logout()

@spec logout() :: Manager.Type.logout_response()

recover_login()

@spec recover_login() :: Manager.Type.recover_login_response()

start_link(deps)

@spec start_link(state()) :: GenServer.on_start()

strategies()

@spec strategies() :: Manager.Type.strategies_response()

syndicates()

@spec syndicates() :: Manager.Type.syndicates_response()