Market Manager v0.1.0 MarketManager.Interpreter

Core of the market manager, where all the logic and communication with outer layers is. Currently, it works more like a bridge between the different ports of the application and manages data between them.

Types

order_request()

Specs

order_request() :: %{
  required(order_type :: String.t()) => String.t(),
  required(item_id :: String.t()) => String.t(),
  required(platinum :: String.t()) => non_neg_integer(),
  required(quantity :: String.t()) => non_neg_integer(),
  required(mod_rank :: String.t()) => non_neg_integer()
}

order_request_without_rank()

Specs

order_request_without_rank() :: %{
  required(order_type :: String.t()) => String.t(),
  required(item_id :: String.t()) => String.t(),
  required(platinum :: String.t()) => non_neg_integer(),
  required(quantity :: String.t()) => non_neg_integer()
}

Functions

activate(syndicate, deps \\ [store: MarketManager.Store.FileSystem, auction_house: MarketManager.AuctionHouse.HTTPClient])

Specs

activate(MarketManager.syndicate(), keyword()) ::
  MarketManager.activate_response()

deactivate(syndicate, deps \\ [store: MarketManager.Store.FileSystem, auction_house: MarketManager.AuctionHouse.HTTPClient])

Specs

deactivate(MarketManager.syndicate(), keyword()) ::
  MarketManager.deactivate_response()