RateLimiter behaviour (Market Manager v2.2.1)

Interface for a rate limiter. Based from: https://akoutmos.com/post/rate-limiting-with-genservers/

Types

metadata()

@type metadata() :: map()

request_arguments()

@type request_arguments() :: [any()]

request_handler()

@type request_handler() :: {(... -> response()), request_arguments()}

response()

@type response() :: any()

response_function()

@type response_function() :: (response(), metadata() -> any())

response_handler()

@type response_handler() :: {response_function(), metadata()}

Callbacks

make_request(request_handler, response_handler)

@callback make_request(request_handler(), response_handler()) :: :ok

Functions

calculate_refresh_rate(num_requests)

Calculates the refresh rate, to determine how many requests per second the limiter can made.

get_rate_limiter()

Gets the limiter from the configs at compile time.

get_requests_per_second()

Gets the requests per second from the configs at compile time.

make_request(request_handler, response_handler)

Default implementation for the make_request function of this behaviour. Fetches the limiter of choice from configs.