Shared.Data.Order (Market Manager v2.1.5)

Represents an order to be made to warframe market. There are two types of orders thus far, RankedOrders and SimpleOrders. See respective files for more information.

order()

@type order() ::
  Shared.Data.Order.RankedOrder.ranked_order()
  | Shared.Data.Order.SimpleOrder.simple_order()

t()

@type t() :: Shared.Data.Order.SimpleOrder.t() | Shared.Data.Order.RankedOrder.t()

new(order)

@spec new(order()) :: t()