Shared.Data.OrderInfo.User (Market Manager v2.2.1)

Represents the account information for a warframe.market User who has an order posted. The current format is the following:

"user": {
  "id": "5962ff05d3ffb64d46e3c47f",
  "ingameName": "JeyciKon",
  "slug": "jeycikon",
  "reputation": 2,
  "platform": "pc",
  "crossplay": true,
  "locale": "pt",
  "status": "ingame",
  "activity": {
    "type": "UNKNOWN",
    "details": "unknown"
  }
}

We only take some of the information, not all. PC Players with crossplay active can trade with any platform (except Nintendo switch). https://www.warframe.com/crossprogression

Types

crossplay()

@type crossplay() :: boolean()

ingame_name()

@type ingame_name() :: String.t()

platform()

@type platform() :: :pc

slug()

@type slug() :: String.t()

status()

@type status() :: :online | :offline | :ingame

t()

@type t() :: %Shared.Data.OrderInfo.User{
  crossplay: crossplay(),
  ingame_name: ingame_name(),
  platform: platform(),
  slug: slug(),
  status: status()
}

Account information of an User

user()

@type user() :: %{
  required(ingameName :: String.t()) => String.t(),
  required(slug :: String.t()) => String.t(),
  required(status :: String.t()) => String.t(),
  required(platform :: String.t()) => String.t(),
  required(crossplay :: String.t()) => boolean()
}

Functions

new(user)

@spec new(user()) :: t()