Shared.Data.Product (Market Manager v2.1.5)

A Product is an item we can create a sell order for. Currently we only support mods. Products contain warframe.market information about the item, as well as default parameters to create sell orders.

product()

@type product() ::
  %{
    required(name :: String.t()) => String.t(),
    required(id :: String.t()) => String.t(),
    required(min_price :: String.t()) => pos_integer(),
    required(default_price :: String.t()) => pos_integer(),
    required(quantity :: String.t()) => pos_integer(),
    required(rank :: String.t()) => non_neg_integer() | String.t()
  }
  | %{
      required(name :: String.t()) => String.t(),
      required(id :: String.t()) => String.t(),
      required(min_price :: String.t()) => pos_integer(),
      required(default_price :: String.t()) => pos_integer()
    }
  | [
      name: String.t(),
      id: String.t(),
      min_price: pos_integer(),
      default_price: pos_integer()
    ]

t()

@type t() :: %Shared.Data.Product{
  default_price: pos_integer(),
  id: String.t(),
  min_price: pos_integer(),
  name: String.t(),
  quantity: pos_integer(),
  rank: non_neg_integer() | String.t()
}

Product details

new(product)

@spec new(product()) :: t()