Store (Market Manager v1.2.0)

Port for the persistency layer. This module contains the Public API for the storage app.

Types

all_orders_store()

Specs

all_orders_store() :: %{
  required(new_loka :: String.t()) => [order_id()],
  required(perrin_sequence :: String.t()) => [order_id()],
  required(red_veil :: String.t()) => [order_id()],
  required(simaris :: String.t()) => [order_id()]
}

delete_order_response()

Specs

delete_order_response() :: {:ok, order_id()} | error()

deps()

Specs

deps() :: keyword()

error()

Specs

error() :: {:error, any()}

list_orders_response()

Specs

list_orders_response() :: {:ok, [order_id()]} | error()

list_products_response()

Specs

list_products_response() :: {:ok, [product()]} | error()

login_info()

Specs

login_info() :: %{
  required(token :: String.t()) => String.t(),
  required(cookie :: String.t()) => String.t()
}

order_id()

Specs

order_id() :: String.t()

product()

Specs

product() :: %{
  required(name :: String.t()) => String.t(),
  required(id :: String.t()) => String.t(),
  required(min_price :: String.t()) => non_neg_integer(),
  required(default_price :: String.t()) => non_neg_integer(),
  required(quantity :: String.t()) => non_neg_integer(),
  required(rank :: String.t()) => non_neg_integer() | String.t()
}

save_order_response()

Specs

save_order_response() :: {:ok, order_id()} | error()

setup_response()

Specs

setup_response() :: {:ok, login_info()} | {:error, :file.posix()}

syndicate()

Specs

syndicate() :: String.t()

syndicate_exists_response()

Specs

syndicate_exists_response() :: {:ok, boolean()} | error()

Functions

delete_order(order_id, syndicate)

Specs

delete_order(order_id(), syndicate()) :: delete_order_response()

Deletes the given orderId from the given syndicate from the storage system.

Example:

> Store.delete_order("00f83ca2-67d9-4019-9fea-587b9fc4037c", "red_veil")
{:ok, "00f83ca2-67d9-4019-9fea-587b9fc4037c"}

> Store.delete_order("invalid_syndicate")
{:error, :enoent}

list_orders(syndicate)

Specs

list_orders(syndicate()) :: list_orders_response()

Lists the ids of currently active orders from the given syndicate.

Example:

> Store.list_orders("red_veil")
{:ok, ["5a750686-956f-42a6-8194-11925ec9281e", "58f0b7e7-0ded-4932-8ccd-380cc5634c82", ...]}

> Store.list_orders("invalid_syndicate")
{:error, :syndicate_not_found}

list_products(syndicate)

Specs

list_products(syndicate()) :: list_products_response()

Lists the products from the given syndicate.

Example:

> Store.list_products("red_veil")
{:ok, [
  %{
    "name" => "Eternal War",
    "id" => "8ee5b3b0-fa43-4dbc-9363-a52930dc742e",
    "min_price" => 14,
    "default_price" => 15,
    "quantity" => 1,
    "rank" => 0
  },
  ...
]}

> Store.list_products("invalid_syndicate")
{:error, :syndicate_not_found}

save_order(order_id, syndicate)

Specs

save_order(order_id(), syndicate()) :: save_order_response()

Saves the given orderId for the given syndicate in the storage system.

Example:

> Store.save_order("00f83ca2-67d9-4019-9fea-587b9fc4037c", "red_veil")
{:ok, "00f83ca2-67d9-4019-9fea-587b9fc4037c"}

> Store.save_order("invalid_syndicate")
{:error, :enoent}

setup(login_info)

Specs

setup(login_info()) :: setup_response()

Saves the setup information from the user into the storage system. Does not perform validation.

Example:

> Store.setup(%{"token" => "a_token", "cookie" => "a_cookie"})
{:ok, %{"token" => "a_token", "cookie" => "a_cookie"}}

> Store.setup(%{"token" => "a_token", "cookie" => "a_cookie"})
{:error, :no_permissions}

syndicate_exists?(syndicate)

Specs

syndicate_exists?(syndicate()) :: {:ok, boolean()} | error()

Returns true if the given syndicate exists, false otherwise. Returns an error if an error occurs.

Example:

> Store.syndicate_exists?(red_veil")
{:ok, true}

> Store.syndicate_exists?("nonexistent_syndicate")
{:ok, false}

> Store.syndicate_exists?("syndicate") # world explodes meanwhile
{:error, :enoent}