WebInterface.Commands (Market Manager v2.1.5)

Contains information about the commands the web_interface understands. Works as a bridge between the web interface and the Manager application, which does the real work.

Types

activate_request()

Specs

activate_request() :: %{
  command: :activate,
  strategy: Manager.Type.strategy(),
  syndicates: [Manager.Type.syndicate()]
}

authenticate_request()

Specs

authenticate_request() :: %{
  command: :authenticate,
  cookie: String.t(),
  token: String.t()
}

command()

Specs

command() :: %{name: String.t(), description: String.t(), id: command_id()}

command_id()

Specs

command_id() :: :activate | :deactivate | :authenticate

deactivate_request()

Specs

deactivate_request() :: %{
  command: :deactivate,
  syndicates: [Manager.Type.syndicate()]
}

dependencies()

Specs

dependencies() :: %{manager: module()}

request()

Specs

request() :: activate_request() | deactivate_request() | authenticate_request()

Functions

execute(command, deps \\ %{manager: Manager})

Specs

execute(request(), dependencies()) :: any()

get_command(id)

Specs

get_command(atom()) :: command() | nil

list_commands()

Specs

list_commands() :: [command()]