WebInterface.CoreComponents (Market Manager v2.1.9)

Provides core UI components.

At the first glance, this module may seem daunting, but its goal is to provide some core building blocks in your application, such modals, tables, and forms. The components are mostly markup and well documented with doc strings and declarative assigns. You may customize and style them in any way you want, based on your application growth and needs.

The default components use Tailwind CSS, a utility-first CSS framework. See the Tailwind CSS documentation to learn how to customize them or feel free to swap in another framework altogether.

Icons are provided by heroicons. See icon/1 for usage.

Functions

back(assigns)

Renders a back navigation link.

Examples

<.back navigate={~p"/posts"}>Back to posts</.back>

Attributes

navigate (:any) (required)

Slots

inner_block (required)

button(assigns)

Renders a button.

Examples

<.button>Send!</.button>
<.button phx-click="go" class="ml-2" disabled=false>Send!</.button>

Attributes

type (:string) - Defaults to nil.

class (:string) - Defaults to nil.

disabled (:boolean) - Defaults to false. Global attributes are accepted.

Slots

inner_block (required)

checkgroup(assigns)

Generate a checkbox group for multi-select.

Examples

<.checkgroup

field={@form[:genres]}
options={[{"Fantasy", "fantasy"}, {"Science Fiction", "sci-fi"}]}
selected={[{"Fantasy", "fantasy"}]}

/>

Attributes

name (:any)

field (Phoenix.HTML.FormField) - a form field struct retrieved from the form, for example: @form[:genres].

required (:boolean) - Defaults to false.

options (:list) - the options to pass to Phoenix.HTML.Form.options_for_select/2. Defaults to [].

disabled (:list) - the list of options that are disabled. Defaults to [].

selected (:list) - the currently selected options, to know which boxes are checked. Defaults to [].

description(assigns)

Renders a description paragraph.

Examples

<.description class="additional-class"> Some description here! </.description>

Attributes

class (:string) - Defaults to nil.

Slots

inner_block (required) - the text representing a description.

error(assigns)

Generates a generic error message.

Slots

inner_block (required)

flash(assigns)

Renders flash notices.

Examples

<.flash kind={:info} flash={@flash} />
<.flash kind={:info} phx-mounted={show("#flash")}>Welcome Back!</.flash>

Attributes

id (:string) - the optional id of flash container. Defaults to "flash".

flash (:map) - the map of flash messages to display. Defaults to %{}.

title (:string) - Defaults to nil.

kind (:atom) - used for styling and flash lookup.Must be one of :info, or :error. Global attributes are accepted.

Slots

inner_block - the optional inner block that renders the flash message.

flash_group(assigns)

Shows the flash group with standard titles and content.

Examples

<.flash_group flash={@flash} />

Attributes

flash (:map) (required) - the map of flash messages.

header(assigns)

Renders a header with title.

Attributes

class (:string) - Defaults to nil.

Slots

inner_block (required)

subtitle

actions

hide(js \\ %JS{}, selector)

hide_dropdown(to)

hide_modal(js \\ %JS{}, id)

icon(assigns)

Renders a Hero Icon.

Hero icons come in three styles – outline, solid, and mini. By default, the outline style is used, but solid an mini may be applied by using the -solid and -mini suffix.

You can customize the size and colors of the icons by setting width, height, and background color classes.

Icons are extracted from your assets/vendor/heroicons directory and bundled within your compiled app.css by the plugin in your assets/tailwind.config.js.

Examples

<.icon name="hero-x-mark-solid" />
<.icon name="hero-arrow-path" class="ml-1 w-3 h-3 animate-spin" />

Attributes

name (:string) (required)

class (:string) - Defaults to nil.

input(assigns)

Renders an input with label and error messages.

A %Phoenix.HTML.Form{} and field name may be passed to the input to build input names and error messages, or all the attributes and errors may be passed explicitly.

Examples

<.input field={@form[:email]} type="email" />
<.input name="my-input" errors={["oh no!"]} />

Attributes

id (:any) - Defaults to nil.

name (:any)

label (:string) - Defaults to nil.

value (:any)

type (:string) - Defaults to "text".

field (Phoenix.HTML.FormField) - a form field struct retrieved from the form, for example: @form[:email].

errors (:list) - Defaults to [].

checked (:boolean) - the checked flag for checkbox inputs.

prompt (:string) - the prompt for select inputs. Defaults to nil.

options (:list) - the options to pass to Phoenix.HTML.Form.options_for_select/2.

multiple (:boolean) - the multiple flag for select inputs. Defaults to false.

selected (:any) - the currently selected option, to know which input(s) is/are currently selected. Defaults to []. Global attributes are accepted.

Slots

inner_block

label(assigns)

Renders a label.

Attributes

for (:string) - Defaults to nil.

Slots

inner_block (required)

list(assigns)

Renders a data list.

Examples

<.list>
  <:item title="Title"><%= @post.title %></:item>
  <:item title="Views"><%= @post.views %></:item>
</.list>

Slots

item (required) - Accepts attributes:

title (:string) (required)

modal(assigns)

Renders a modal.

Examples

<.modal id="confirm-modal">
  This is a modal.
</.modal>

JS commands may be passed to the :on_cancel to configure the closing/cancel event, for example:

<.modal id="confirm" on_cancel={JS.navigate(~p"/posts")}>
  This is another modal.
</.modal>

Attributes

id (:string) (required)

show (:boolean) - Defaults to false.

on_cancel (Phoenix.LiveView.JS) - Defaults to %Phoenix.LiveView.JS{ops: []}.

Slots

inner_block (required)

progress_bar(assigns)

Renders a progress bar for an ongoing operation.

Examples

<.progress_bar hidden=false progress=15 />
<.progress_bar hidden=true progress=0 message="Activating system ..." />
<.progress_bar class="cool-bar" />

Attributes

hidden (:boolean) - whether or not to show the progress bar. Defaults to false.

progress (:integer) - the current progress of the bar. Defaults to 0.

message (:string) - the message to show while the bar is progressing. Defaults to "Operation in progress ...".

class (:string) - Defaults to nil.

radiogroup(assigns)

Generate a radio button group.

Examples

<.radiogroup

field={@form[:genres]}
label="Genres"
options={[{"Fantasy", "fantasy"}, {"Science Fiction", "sci-fi"}]}
selected={%{"Science Fiction", "sci-fi"}}

/>

Attributes

id (:any)

name (:any)

label (:string) - Defaults to nil.

field (Phoenix.HTML.FormField) - a form field struct retrieved from the form, for example: @form[:genres].

errors (:list)

required (:boolean) - Defaults to false.

options (:list) - the options to pass to Phoenix.HTML.Form.options_for_select/2.

class (:string) - Defaults to nil.

selected (:any) - the currently selected option, to know which radio button is checked. Defaults to nil. Global attributes are accepted.

section(assigns)

Renders the title and description for a new section.

Examples

<.section title="Strategy" description="This is a cool strategy!" />

Attributes

title (:string) (required) - the title of the section.

description (:string) (required) - the description of the section.

show(js \\ %JS{}, selector)

show_modal(js \\ %JS{}, id)

simple_form(assigns)

Renders a simple form.

Examples

<.simple_form for={@form} phx-change="validate" phx-submit="save">
  <.input field={@form[:email]} label="Email"/>
  <.input field={@form[:username]} label="Username" />
  <:actions>
    <.button>Save</.button>
  </:actions>
</.simple_form>

Attributes

for (:any) (required) - the data-structure for the form.

as (:any) - the server side parameter to collect all input under. Defaults to nil. Global attributes are accepted.

Slots

inner_block (required)

actions - the slot for form actions, such as a submit button.

table(assigns)

Renders a table with generic styling.

Examples

<.table id="users" rows={@users}>
  <:col :let={user} label="id"><%= user.id %></:col>
  <:col :let={user} label="username"><%= user.username %></:col>
</.table>

Attributes

id (:string) (required)

rows (:list) (required)

row_id (:any) - the function for generating the row id. Defaults to nil.

row_click (:any) - the function for handling phx-click on each row. Defaults to nil.

row_item (:any) - the function for mapping each row before calling the :col and :action slots. Defaults to &Function.identity/1.

Slots

col (required) - Accepts attributes:

label (:string)

action - the slot for showing user actions in the last table column.

toggle_dropdown(to)

translate_error(arg)

Translates an error message using gettext.

translate_errors(errors, field)

Translates the errors for a field from a keyword list of errors.