WebInterface.Syndicates (Market Manager v1.2.0)

Contains information about the available syndicates that Manager supports as well as utility functions to work with them.

Types

syndicate_info()

Specs

syndicate_info() :: %{
  required(name :: atom()) => String.t(),
  required(id :: atom()) => Manager.syndicate()
}

Functions

get_id(syndicate)

Specs

get_id(syndicate_info()) :: Manager.syndicate() | nil

get_syndicate(id)

Specs

get_syndicate(String.t()) :: syndicate_info() | nil

list_syndicates()

Specs

list_syndicates() :: [syndicate_info()]