AuctionHouse (Market Manager v2.2.1)

Library representing the interface for the auction house. Responsible for making calls and decoding the answers from the auction house into a format the manager understands.

Functions

delete_order(order)

@spec delete_order(Shared.Data.PlacedOrder.t()) ::
  AuctionHouse.Type.delete_order_response()

Deletes an order from warframe market asynchronously. Notifies invoking process asynchronously with placed order or with error.

Example:

> alias Shared.Data.PlacedOrder
> placed_order = PlacedOrder.new(%{
    "item_id" => "54e644ffe779897594fa68cd",
    "order_id" => "626127cbc984ac033cd2bbd2"
})

> AuctionHouse.delete_order(placed_order)
:ok

The received messages will one of the following formats:

- {:delete_order, {:ok, %PlacedOrder{item_id: "54e644ffe779897594fa68cd", order_id: "626127cbc984ac033cd2bbd2"}}}
- {:delete_order, {:error, {reason, err}}}

get_item_orders(item_name)

@spec get_item_orders(AuctionHouse.Type.item_name()) ::
  AuctionHouse.Type.get_item_orders_response()

Gets all warframe market orders for the item with the given name. The item's name is in human readable format. This function also converts the name into a format that the external party can understand. Notifies invoking process asynchronously with a list of OrderInfo or with error.

Example:

> alias Shared.Data.OrderInfo
> item_name = "Despoil"

> AuctionHouse.get_item_orders(item_name)
> :ok

The received messages will one of the following formats:

- {:get_item_orders,  {:ok, "Despoil", [%OrderInfo{"visible" => true, "order_type" => "sell", "platform" => "pc",
  "platinum" => 20, "user" => %OrderInfo.User{"ingame_name" => "user_name_1", "status" => "ingame"}}]}
- {:get_item_orders, {:error, "Despoil", {reason, err}}}

get_saved_login()

@spec get_saved_login() :: AuctionHouse.Type.get_saved_login_response()

Returns the authorization and user directly from state. Synchronous call, will only fail if this service is down.

Example:

> alias Shared.Data.{Authorization, User}
> auth = Authorization.new("a_cookie", "a_token")
> user = User.new("fl4m3", false)

> AuctionHouse.get_saved_login()
{:ok, {%Authorization{}, %User{}}}

> AuctionHouse.get_saved_login()
{:error, :reason}

get_user_orders(username_slug)

@spec get_user_orders(AuctionHouse.Type.username_slug()) ::
  AuctionHouse.Type.get_user_orders_response()

Gets all warframe market sell orders for the given user. Notifies invoking process asynchronously with a list of user PlacedOrders or with error.

Example:

> AuctionHouse.get_user_orders("fl4m3")
> :ok

The received messages will one of the following formats:

- {:get_user_orders, {:ok,
  [
    %PlacedOrder{
      order_id: "66058313a9630600302d4889",
      item_id: "55108594e77989728d5100c6"
    },
    %PlacedOrder{
      order_id: "6605832ea96306003657a90d",
      item_id: "54e644ffe779897594fa68d2"
    }
  ]
}}

- {:get_user_orders, {:error, {reason, err}}}

login(credentials)

@spec login(Shared.Data.Credentials.t()) :: AuctionHouse.Type.login_response()

Authenticates the user with the auction house and saves the session. Must be invoked every time the application is launched. Notifies invoking process asynchronously with user information or with error.

Example:

> alias Shared.Data.{Authorization, Credentials, User}
> credentials = Credentials.new("the_username", "the_password")

> AuctionHouse.login(credentials)
:ok

The received messages will one of the following formats:

- {:login, {:ok, { %Authorization{cookie: "a_cookie", token: "a_token"}, %User{patreon?: false, slug: "f4lm3", ingame_name: "Fl4m3"}}}}
- {:login, {:error, {reason, err}}}

logout()

@spec logout() :: AuctionHouse.Type.logout_response()

Deletes the current session and user data from the this application. Does not interact with the External AuctionHouse, this is a local operation only. Will only fail if this service is down.

Example:

> AuctionHouse.logout()
:ok

place_order(order)

@spec place_order(Shared.Data.Order.t()) :: AuctionHouse.Type.place_order_response()

Places an order in warframe market. Notifies invoking process asynchronously with placed order or with error.

Example:

> alias Shared.Data.{Order, PlacedOrder}
> order = Order.new(%{
  "item_id" => "54e644ffe779897594fa68cd",
  "mod_rank" => 0,
  "order_type" => "sell",
  "platinum" => 20,
  "quantity" => 1
})

> AuctionHouse.place_order(order)
:ok

The received messages will have one of the following formats:

- {:place_order, {:ok, %PlacedOrder{item_id: "54a74454e779892d5e5155d5", order_id: "66b9c7aa6b17410a57974e4b"}}}
- {:place_order, {:error, {reason, err}}}

update_login(auth, user)

@spec update_login(Shared.Data.Authorization.t(), Shared.Data.User.t()) ::
  AuctionHouse.Type.update_login_response()

Feeds the authorization information directly to the AuctionHouse. Used when the login data is being recovered from a past login to update the this service's state. Synchronous call, will only fail if this service is down.

Example:

> alias Shared.Data.{Authorization, User}
> auth = Authorization.new("a_cookie", "a_token")
> user = User.new("fl4m3", false)

> AuctionHouse.update_login(auth, user)
:ok