Market Manager v1.0.0 Manager

MarketManager is an application that allows you to make batch requests to warframe.market. This is the entrypoint of everything. If you have a module and you need to talk to MarketManager, this is who you call, the public API.

Types

activate_response()

Specs

activate_response() ::
  {:ok, :success}
  | {:partial_success, [{:error, error_reason(), item_id()}, ...]}
  | {:error, :unable_to_place_requests, [{:error, error_reason(), item_id()}]}

deactivate_response()

Specs

deactivate_response() ::
  {:ok, :success}
  | {:partial_success, [{:error, error_reason(), order_id()}, ...]}
  | {:error, :unable_to_delete_orders, [{:error, error_reason(), order_id()}]}

error_reason()

Specs

error_reason() :: atom()

item_id()

Specs

item_id() :: String.t()

order_id()

Specs

order_id() :: String.t()

strategy()

Specs

strategy() :: atom()

syndicate()

Specs

syndicate() :: String.t()

Functions

activate(syndicate, strategy)

Specs

activate(syndicate(), strategy()) :: activate_response()

Activates a syndicate in warframe.market. Activating a syndicate means you put on sell all the mods the syndicate has with that are in the products.json file. The price of each mod will be calculated via a PriceAnalyst depending on <<<<<<< HEAD which strategy you choose.

Example:

{:ok, :success} = MarketManager.activate("simaris", :lowest_minus_one)

======= which strategy you choose.
Example:

{:ok, :success} = MarketManager.activate("simaris", :lowest_minus_one)

>>>>>>> master

deactivate(syndicate)

Specs

deactivate(syndicate()) :: deactivate_response()

Deactivates a syndicate in warframe.market. Deactivating a syndicate means you <<<<<<< HEAD delete all orders you have placed before that belong to the given syndicate.

Example:

{:ok, :success} = MarketManager.deactivate("simaris")

======= delete all orders you have placed before that belong to the given syndicate.
Example:

{:ok, :success} = MarketManager.deactivate("simaris")

>>>>>>> master

valid_strategy?(strategy)

Specs

valid_strategy?(String.t()) :: boolean()

<<<<<<< HEAD
Returns true if the given strategy is valid, false otherwise.

Example:

MarketManager.valid_strategy?("bananas")          # false
MarketManager.valid_strategy?("equal_to_lowest")  # true

=======
Returns true if the given strategy is valid, false otherwise.

Example:

MarketManager.valid_strategy?("bananas")          # false
MarketManager.valid_strategy?("equal_to_lowest")  # true

>>>>>>> master