Market Manager v1.0.0 Manager.Interpreter

Core of the manager, where all the logic and communication with outer layers is. Currently, it works more like a bridge between the different ports of the application and manages data between them.

Types

order_request()

Specs

order_request() :: %{
  required(order_type :: String.t()) => String.t(),
  required(item_id :: String.t()) => String.t(),
  required(platinum :: String.t()) => non_neg_integer(),
  required(quantity :: String.t()) => non_neg_integer(),
  required(mod_rank :: String.t()) => non_neg_integer()
}

order_request_without_rank()

Specs

order_request_without_rank() :: %{
  required(order_type :: String.t()) => String.t(),
  required(item_id :: String.t()) => String.t(),
  required(platinum :: String.t()) => non_neg_integer(),
  required(quantity :: String.t()) => non_neg_integer()
}

Functions

activate(syndicate, strategy, deps \\ [store: Store, auction_house: AuctionHouse])

Specs

activate(Manager.syndicate(), Manager.strategy(), keyword()) ::
  Manager.activate_response()

deactivate(syndicate, deps \\ [store: Store, auction_house: AuctionHouse])

Specs

deactivate(Manager.syndicate(), keyword()) :: Manager.deactivate_response()

valid_action?(action)

Specs

valid_action?(String.t()) :: boolean()