Manager.Runtime.Worker (Market Manager v2.1.5)

Process responsible for doing asynchrnous tasks. The Worker is supervised by the Server and is restarted should something fail. It communicates with the calling client via send/2, since we don't know if the client is a GenServer or not. Should this change (and we are sure the client is a GenServer) then GenServer.cast should be used instead.

Types

from()

Specs

from() :: pid()

state()

Specs

state() :: keyword(module())

Functions

activate(syndicate, strategy)

Specs

activate(Manager.Type.syndicate(), Manager.Type.strategy()) :: :ok

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

deactivate(syndicate)

Specs

deactivate(Manager.Type.syndicate()) :: :ok

start_link(deps)

Specs

start_link(state()) :: GenServer.on_start()