Store (Market Manager v2.1.7)

Port for the persistency layer. This module contains the Public API for the storage app.

Functions

delete_login_data()

@spec delete_login_data() :: Store.Type.delete_login_data_response()

Deletes any login information from the storage system.

Example:

> Store.delete_login_data()
:ok

> Store.delete_login_data()
{:error, :enoent}

delete_order(placed_order, syndicate)

@spec delete_order(Shared.Data.PlacedOrder.t(), Shared.Data.Syndicate.t()) ::
  Store.Type.delete_order_response()

Deletes the given placed_order from the given syndicate from the storage system.

Example:

> alias Shared.Data.{PlacedOrder, Syndicate}

> Store.delete_order(
  %PlacedOrder{item_name: "Exothermic", order_id: "5526aec1e779896af9418266"},
  Syndicate.new(name: "Red Veil", id: :red_veil, catalog: [])
)
:ok

> invalid_syndicate = Syndicate.new(name: "Bad Synd", id: :bad_syndicate, catalog: [])
> Store.delete_order(invalid_syndicate)
{:error, :enoent}

get_login_data()

@spec get_login_data() :: Store.Type.get_login_data_response()

Retrieves the user's login data from Storage.

Example:

> alias Shared.Data.{Authorization, User}

> Store.get_login_data()
{:ok, {
  %Authorization{token: "a_token", cookie: "a_cookie"},
  %User{ingame_name: "username", patreon?: false}
}}

> Store.get_login_data()
{:error, :enoent}

list_active_syndicates()

@spec list_active_syndicates() :: Store.Type.list_active_syndicates_response()

Returns the syndicates currently active. An active syndicate is a syndicate with orders in current_orders.

Example:

> alias Shared.Data.Syndicate

> Store.list_active_syndicates()
{:ok, [%Syndicate{name: "Red Veil", id: :red_veil, catalog: []}, %Syndicate{name: "New Loka", id: :new_loka, catalog: []}]}

> Store.list_active_syndicates()
{:error, :enoent}

list_orders(syndicate)

@spec list_orders(Shared.Data.Syndicate.t()) :: Store.Type.list_orders_response()

Lists all placed orders from the given syndicate.

Example:

> alias Shared.Data.{PlacedOrder, Syndicate}

> syndicate = Syndicate.new(name: "Red Veil", id: :red_veil, catalog: [])
> Store.list_orders(syndicate)
{:ok, [
  %PlacedOrder{item_name: "Exothermic", order_id: "5526aec1e779896af9418266"},
  %PlacedOrder{item_name: "Tribunal", order_id: "5ea087d1c160d001303f9ed7"},
  ...
]}

> invalid_syndicate = Syndicate.new(name: "Bad Synd", id: :bad_syndicate)
> Store.list_orders(invalid_syndicate)
{:error, :syndicate_not_found}

list_products(syndicate)

@spec list_products(Shared.Data.Syndicate.t()) :: Store.Type.list_products_response()

Lists the products from the given syndicate.

Example:

> alias Shared.Data.{Product, Syndicate}

> syndicate = Syndicate.new(name: "Red Veil", id: :red_veil, catalog: [])
> Store.list_products(syndicate)
{:ok, [
  %Product{
    name: "Eternal War",
    id: "8ee5b3b0-fa43-4dbc-9363-a52930dc742e",
    min_price: 14,
    default_price: 15,
    quantity: 1,
    rank: 0
  },
  ...
]}

> invalid_syndicate = Syndicate.new(name: "Bad Synd", id: :bad_syndicate, catalog: [])
> Store.list_products(invalid_syndicate)
{:error, :syndicate_not_found}

list_syndicates()

@spec list_syndicates() :: Store.Type.list_syndicates_response()

Returns all the syndicates currently stored. The fact a syndicate is stored does not mean it has products, nor orders.

Example:

> alias Shared.Data.Syndicate

> Store.list_syndicates()
{:ok, [%Syndicate{name: "Red Veil", id: :red_veil, catalog: []}, %Syndicate{name: "New Loka", id: :new_loka, catalog: []}]}

> Store.list_syndicates()
{:error, :enoent}

save_login_data(auth, user)

@spec save_login_data(Shared.Data.Authorization.t(), Shared.Data.User.t()) ::
  Store.Type.save_login_data_response()

Saves the login information from the user into the storage system. Does not perform validation.

Example:

> alias Shared.Data.{Authorization, User}

> Store.save_login_data(
  %Authorization{token: "a_token", cookie: "a_cookie"},
  %User{ingame_name: "username", patreon?: false}
)
:ok

> Store.save_login_data(
  %Authorization{token: "a_token", cookie: "a_cookie"},
  %User{ingame_name: "username", patreon?: false}
)
{:error, :enoent}

save_order(placed_order, syndicate)

@spec save_order(Shared.Data.PlacedOrder.t(), Shared.Data.Syndicate.t()) ::
  Store.Type.save_order_response()

Saves the given placed_order for the given syndicate in the storage system.

Example:

> alias Shared.Data.{PlacedOrder, Syndicate}

> Store.save_order(
  %PlacedOrder{item_name: "Exothermic", order_id: "5526aec1e779896af9418266"},
  Syndicate.new(name: "Red Veil", id: :red_veil, catalog: [])
)
:ok

> invalid_syndicate = Syndicate.new(name: "Bad Synd", id: :bad_syndicate, catalog: [])
> Store.save_order(invalid_syndicate)
{:error, :enoent}