Store (Market Manager v2.1.3)

Port for the persistency layer. This module contains the Public API for the storage app.

Functions

delete_order(order_id, syndicate)

Specs

delete_order(Store.Type.order_id(), Store.Type.syndicate()) ::
  Store.Type.delete_order_response()

Deletes the given orderId from the given syndicate from the storage system.

Example:

> Store.delete_order("00f83ca2-67d9-4019-9fea-587b9fc4037c", "red_veil")
{:ok, "00f83ca2-67d9-4019-9fea-587b9fc4037c"}

> Store.delete_order("invalid_syndicate")
{:error, :enoent}

get_credentials()

Specs

get_credentials() :: Store.Type.get_credentials_response()

Retrieves the user's authentication from Storage.

Example:

> Store.get_credentials()
{:ok, %{"token" => "a_token", "cookie" => "a_cookie"}}

> Store.get_credentials()
{:error, :enonent}

list_orders(syndicate)

Specs

list_orders(Store.Type.syndicate()) :: Store.Type.list_orders_response()

Lists the ids of currently active orders from the given syndicate.

Example:

> Store.list_orders("red_veil")
{:ok, ["5a750686-956f-42a6-8194-11925ec9281e", "58f0b7e7-0ded-4932-8ccd-380cc5634c82", ...]}

> Store.list_orders("invalid_syndicate")
{:error, :syndicate_not_found}

list_products(syndicate)

Specs

list_products(Store.Type.syndicate()) :: Store.Type.list_products_response()

Lists the products from the given syndicate.

Example:

> Store.list_products("red_veil")
{:ok, [
  %{
    "name" => "Eternal War",
    "id" => "8ee5b3b0-fa43-4dbc-9363-a52930dc742e",
    "min_price" => 14,
    "default_price" => 15,
    "quantity" => 1,
    "rank" => 0
  },
  ...
]}

> Store.list_products("invalid_syndicate")
{:error, :syndicate_not_found}

save_credentials(login_info)

Specs

save_credentials(Store.Type.login_info()) ::
  Store.Type.save_credentials_response()

Saves the authentication information from the user into the storage system. Does not perform validation.

Example:

> Store.save_credentials(%{"token" => "a_token", "cookie" => "a_cookie"})
{:ok, %{"token" => "a_token", "cookie" => "a_cookie"}}

> Store.save_credentials(%{"token" => "a_token", "cookie" => "a_cookie"})
{:error, :no_permissions}

save_order(order_id, syndicate)

Specs

save_order(Store.Type.order_id(), Store.Type.syndicate()) ::
  Store.Type.save_order_response()

Saves the given orderId for the given syndicate in the storage system.

Example:

> Store.save_order("00f83ca2-67d9-4019-9fea-587b9fc4037c", "red_veil")
{:ok, "00f83ca2-67d9-4019-9fea-587b9fc4037c"}

> Store.save_order("invalid_syndicate")
{:error, :enoent}

syndicate_exists?(syndicate)

Specs

syndicate_exists?(Store.Type.syndicate()) ::
  {:ok, boolean()} | Store.Type.error()

Returns true if the given syndicate exists, false otherwise. Returns an error if an error occurs.

Example:

> Store.syndicate_exists?(red_veil")
{:ok, true}

> Store.syndicate_exists?("nonexistent_syndicate")
{:ok, false}

> Store.syndicate_exists?("syndicate") # world explodes meanwhile
{:error, :enoent}